Remove prior support for cached DynamicObject members
After a great deal of trying to cache members, not in the way this code
was implemented, but using a better devised Func<object, object>
handler, and re-enabling the cache and disabling the dynamic lookups, I
ran some tests against the PropertyDictionary. While it worked for a
single object and for the updated property value and it worked for new
objects with the same properties, it failed when new objects contained
different fields and not the original fields when derived from the same
DynamicObject type. While I could have written it in such a way that it
removes the missing members from the cache, I have reason to believe
that the cache introduces issues for multithreaded applications when
used with DynamicObject. I didn't test it, but I could see a parallel
task creating problems when the objects are derived from the same shared
type and are shaped with uniquely different properties. Support removed
for caching DynamicObject objects for type safety. DynamicObject objects
are, by definition, slow, anyways.