Requiring Dynamitey dependency again, broken support, no test exists that predicts misbehavior
Added Dynamitey back in order to fix a problem with a project that
depends on this fork of Mustache#. There are not any tests that
currently exist that can demonstrate the misbehavior that occurs when
incorrectly invoking a `TryGetMember()` implementation. But, Dynamitey
is spot on. They put a lot of effort into being able to invoke a member
of a `DynamicObject`. As far as I'm concerned, there is no need to
duplicate this effort.